little re-org
